Very close to the finish
added predicate for matFunc at the end name equallySized to check sizes
of matrices (it wasn't a good idea to think it wasn't needed), added
tests for it, finished project report (still not here), added
demonstration.clj - a clojure file that contains instructions for REPL
run mode (in case you do not want to use the GUI) with templates for
matrices  and functions to use easily, and some more insights into the
project. Also, added screenshots folder with screenshots showing many
right (and some wrong) ways to use the GUI, with output. To finish,
some more documentation is needed, making some functions more readable,
finishing the report, and maybe adding some more stuff to the
demonstration. Other than that, this is can be considered as a final
version.